Changes from self-review
