+++ java-client/session/src/test/java/io/deephaven/client/impl/UpdateBySpecBuilderTest.java
@@ -351,7 +351,6 @@ public class UpdateBySpecBuilderTest {
         }
     }
 
-
     @Test
     void ema() {
         check(EmaSpec.ofTime("Timestamp", Duration.ofNanos(1)));
